Feat/adminstrative api (#22)
* feat: Add secure API endpoint to populate products

* fix: check if user authenticated or not before shoing products

* feat: Remove default path from populate_products command

This commit removes the hardcoded default JSON file path from the management command. This ensures the command uses the path provided via the API request, preventing "file not found" errors when using a different JSON file.

* fix: Correct URL routing for populate_db endpoint

This commit adjusts the URL patterns to correctly route POST requests to the `populate_db` endpoint. The specific path is now defined before the general router patterns to prevent it from being incorrectly caught, resolving the "405 Method Not Allowed" error.

* fix: Pass json_path from API request to management command

This commit fixes a bug in the `PopulateProductsView` where the `json_path` was not being passed from the API request payload to the `call_command` function. This change ensures the correct JSON file is used to populate the database.

* feat: Add a script to automate database population

This commit adds a new shell script to easily trigger the `populate_db` management command via a cURL request. The script is configured to use command-line arguments for the seller name and JSON file path.

* fix: remove token